Major work towards the AMD64 port.  Most important of which is that continuations work, more assembly instructions are testing in 64-bit mode, and some arithmetic operations (shift-right, etc.) have been fixed by removing some 32-bit dependencies.

+++ src/ikarus-exec.c
@@ -25,7 +25,7 @@
 typedef struct {
   ikptr tag;
   ikptr top;
-  int size;
+  long int size;
   ikptr next;
 } cont;
